Optionally delay audio track selection.
Defaults to delaying by 500ms. Can be modified via KPPlayerConfig.
This helps with initial audio selection issue.

+++ KALTURAPlayerSDK/KPViewControllerProtocols.h
@@ -76,7 +76,7 @@ typedef enum{
 - (void)changeSubtitleLanguage:(NSString *)languageCode;
 - (void)setSourceWithAsset:(AVURLAsset*)asset;
 - (void)hidePlayer;
-- (void)selectAudioTrack:(int)trackId;
+- (void)selectAudioTrack:(NSUInteger)trackId;
 - (void)selectTextTrack:(NSString *)locale;
